refactor(genai-util): organize tests separately
